requirements.txt
Flask
Flask-Login
pyodbc
python-dotenv
pandas
SQLAlchemy
